BF: update: Abort if updating a dataset on an unborn branch
If update() is called while on an unborn branch, trying to merge in
changes will fail with a confusing "not something we can merge" error
from git, relayed by git-annex (see gh-4347).  This isn't a situation
a caller is likely to be in knowingly, but it can happen easily if a
dataset was cloned from a repo where HEAD points to an unborn branch.
That, in turn, can happen easily because create_sibling(), when called
with a branch other than master checked out, will leave HEAD pointing
to master, which may not exist and, even it does, may never be
published to the sibling (gh-4349).

Note that this is a maint-specific fix.  On master update() fails with
a reasonable "cannot determine merge target" (*), a change that came
with 2342cb46e0 (Merge pull request #4167 from kyleam/update-follow,
2020-02-21).

Fixes gh-4347, in the sense that a more understandable error is given.
The situation will be further improved in the rest of this series by
(1) teaching create_sibling() to point the sibling's HEAD to the ref
that is checked out in the local repo and (2) teaching clone() and
friends to try to check out another ref if the one pointed at by HEAD
doesn't exist.

(*) Unsurprisingly, this commit has non-trivial conflicts with master.