ENH: clone: Try to switch away from an unborn branch
The last commit made update() fail with a clearer error message when
called on a branch with no commits.  As mentioned there, a likely
reason for the caller to get into this situation is by cloning a
remote whose HEAD points to a non-existent ref.

Let's try to avoid this scenario by looking for another ref to check
out if we land on an unborn branch.  Select the ref with the most
recent commit, ignoring the git-annex branch.  As mentioned in gh-4349
(*), we could do a smarter check in the case of submodules, but that
would require a more substantial rework that might not be worth the
effort given that, as described in the code comment, the worst case
scenario is a detached HEAD.  (And even if we want to go that
direction eventually, ending up on a detached HEAD is certainly an
improvement over crashing.)

Re: #4349
(*): https://github.com/datalad/datalad/issues/4349#issuecomment-605453854