ENH: create-sibling: Set HEAD of sibling to local repo's branch
When a caller creates a sibling with a branch other than master
checked out, the sibling's HEAD ref still points to refs/heads/master.
Depending on what is pushed to the repo, this may stay an unborn
branch.  With a plain 'git clone', there will be nothing to check out,
so the cloned repo will end up an unborn master.

Earlier commits in this series improved datalad's handling of such a
sibling: `datalad update` fails with a more informative error (as it
already did on master), GitRepo.update_submodule() fails with an
informative error, and `datalad clone` and friends try to check out
another ref when the repo lands on an unborn branch.

While create-sibling's behavior here is less problematic after the
other changes in this series, it can still lead to issues when
datasets are consumed with a plain git clone (or with older versions
of datalad).  Mitigate these remaining issues by making create-sibling
set the sibling's HEAD to the current local branch, which probably
matches the expectations of most callers anyway.

Note: The get_corresponding_branch() call is written in an unusual way
for maint; it's done for compatibility with the master branch, where
get_corresponding_branch() now returns None rather than the main
branch when not on an adjusted branch.

Supersedes and closes #4355.
Closes #4349.