ENH: annexrepo: Set annex.retry=3 when calling git-annex
git-annex has an annex.retry option that instructs it to retry on
transfer failure, but it defaults to 0.  Add a datalad.annex.retry
option that defaults to 3 and is used as the value for annex.retry in
our calls to git-annex, taking care to not override annex.retry if the
user has configured it.

This implements @yarikoptic's proposal from gh-2896 and will hopefully
help resolve some flaky test failures (e.g., gh-3653 and gh-4371).

Closes #2896.