postclone reconfiguration for ORA remote
Idea:
Whenever we clone from a ria+ URL and we find no autoenabled ORA remote
after git-annex-init, we look up the special remote uuid that should be
stored in the store's config, check whether we have that and try
enabling it with the store URL we cloned from.
Rational: The special remote could have a URL stored, that isn't valid
from the point of view of whoever clones from the store for a lack of
credentials or URL resolution or whatever. Autoenable would then fail.
But the way we cloned the repo in the first place seems a promising
candidate to try with the special remote as well.

Note, that in case we cloned via HTTP, we just try to enable origin.
This should work for regular annex keys, but not for an archive as
managed by ORA remote.