TST: annexrepo: Don't assume initial number of git-annex commits
What we care about is that, depending on the value of alwayscommit, we
see the expected _change_ in the number of commits on the git-annex
branch.  Putting in an assumption about the number of initial commits
doesn't improve the check and is prone to breakage if things change on
git-annex's side.