RF: Make default number of annex jobs configurable and 1 by default
The previous default behavior `(max(8, min(3, nCPUs)))` could cause
issues (crashes, system stalls) due to the large number of child
processes spawned and file descriptors opened.

This change introduces a new config switch
`datalad.runtime.max-annex-jobs` that can be used to tailor
parallelization behavior, and effectively replaces the previously
hard-coded maximum value of `8`. Moreover, it drops the default to `1`
to disable parallel annex jobs by default to ensure a potentially
slower, but more robust behavior out of the box.

The config switch is specific to annex jobs in order to leave room for
an anticipated parallelization of some functionality in DataLad itself.
This will likely need separate configuration to be able to handle the
multiplicative impact of parallelization on two levels.

Implementation detail: The number of jobs is now lazily evaluated on
first use. While it only takes ~100ns to get the CPU count, adding a
config query would make this longer (still microsecond range). However,
the configuration scope would then be limited to the user-level, and
could not be tailored to individual datasets, which may be desired for
datasets with a huge number of files.

I left the previously defined `N_AUTO_JOBS` symbol in place to avoid
breaking consumer code (I am not aware of any). But it now also reflects
the changed default behavior.

Fixes gh-4404
Fixes gh-4141