TST: clone: Tweak unborn head tests for adjusted branch compatibility
Two of the tests added in 8f5cc450c2 (ENH: clone: Try to switch away
from an unborn branch, 2020-04-01) fail on the crippled file system
run [0].  The easy part to fix is to write assertions about the
current branch using get_corresponding_branch (written a bit oddly for
compatibility with the changed return value in master).

The more involved part is changing the setup in test_clone_unborn_head
to account for the fact that, while setting up, we may be on an
adjusted branch.  This means we can't use empty commits, because
git-annex diffs to decide whether things need to be propagated back.
And we need to sprinkle in calls to adjust() followed by save() to
sync state.

[0]: Example:
     https://github.com/datalad/datalad/pull/4405/checks?check_run_id=592761391

Re: #4370