RF: <Dataset path=...> -> Dataset('<path>')
__repr__ of a dataset always uses the 'path=' prefix to the path, but a
dataset is also always and only identified by a path. Hence the prefix
is superfluous, and only takes space. On any platform an absolute path
is easy to recognize as such.

This change switched the formatting to yield a C&P ready path, with
all needed quoting for the cmdline.