ENH: interface: Give per-result feedback on suppressed messages
Suppressing messages can lead to a confusing silence if there are lots
of suppressed messages _and_ the underlying operation takes time.  To
let the user know that something is going on, let's write

    N similar messages have been suppressed

each time we receive the result, updating it in place and ending in
the same place as we did before.

Closes #4413.