RF: Demote symlinked CWD related log message to debug-level
This is the message:

[WARNING] realpath of PWD=</some/path/to/symlinked/file> is </resolved/path/to/symlinked/file> whenever os.getcwd()=<path/to/dataset>

Questions about it have motivated a dedicated "please ignore it" PR
in the handbook: https://github.com/datalad-handbook/book/pull/438

I think this message is pointless (but still unconditionally shown) if
everything works. When something doesn't work, it is likely caused by
a bug in DataLad. Hence there is no need to bother user with this
warning that they can or should do nothing about.