ENH+RF: use repr, not quote_cmdlinearg for repr, all *Repos need a single __repr__ implementation
With quote_cmdlinearg we would get inconsistent reprs -- quotes will be used
only if needed, thus it would be less of a true repr and imho it is suboptimal
to have such a jumpy appearance. e.g. in master we would have

	$> python -c 'from datalad.api import Dataset; ds = Dataset("/tmp/v e\" ры"); print(ds, ds.repo)'
	Dataset('/tmp/v e" ры') AnnexRepo('/tmp/v e" ры')
	$> python -c 'from datalad.api import Dataset; ds = Dataset("///"); print(ds, ds.repo)'
	Dataset(/home/yoh/datalad) AnnexRepo(/home/yoh/datalad)

with the changes here it would be more consistent imho:

	$> python -c 'from datalad.api import Dataset; ds = Dataset("/tmp/v e\" ры"); print(ds, ds.repo)'
	Dataset('/tmp/v e" ры') AnnexRepo('/tmp/v e" ры')
	$> python -c 'from datalad.api import Dataset; ds = Dataset("///"); print(ds, ds.repo)'
	Dataset('/home/yoh/datalad') AnnexRepo('/home/yoh/datalad')

Also it is really not needed to duplicate code in __repr__s of *Repos.
Better imho to rely on inheritance and knowing the class name.