RF: Rope all sanity checks into the source loop, use `dataset` as a switch to save
Given that it supports reading arbitrary source/dest pairs from `stdin`,
there is no way to check sanity upfront.

The `dataset` arg was previously semi-unnecessary, because it would
determine and save a single target dataset from `target_dir`. However,
I think it is more useful to flexibly support eating any src/dest pairs
via `specs-from`, rather then forcing users to specify a `target_dir`
that would than (likely) have to be interpreted as the base dir for
relative destination paths. When feeding info from other tools via
`stdin` that seems like a needless complication. Instead, whether or not
to save is now decided based on giving a `dataset` argument. Sanity
checking (if something is actually copied into the to-be-saved dataset
is still done in-loop, and would lead to error results if not.