ENH: Support dataset aliases in RIA stores
`clone()` can now handle URLs of the format

   'ria+<protocol>://<storelocation>#~<aliasname>'

in addition to those containing UUIDs.

A future addition should likely equip `create-sibling-ria()` with the ability
to establish those aliases too. But for now it is not clear to me how to
do that best (in particular how to perform properly on reconfigure).
Given time-pressure this needs to wait for another time.

Fixes gh-4424