BF: cmdline: Restore argcomplete functionality
The optimization in fb38c7098 (OPT/RF: create only the subparser
needed for the command, offer "closest" choices, 2018-05-17) sped up
datalad, but it (would have [*]) broken the argcomplete-based
command-line completion due to the spots in setup_parser() that raise
an exception rather than returning the parser that argcomplete needs
to inspect.

To make completion work again, add a flag to setup_parser() that
signals that datalad is being called by argcomplete.  In this
scenario, avoid failing early so that argcomplete can do its thing.

Closes #4473.

[*]: When I tried from fb38c7098^, completion still didn't work
     because an exception is raised in config.py due to cfg_kv_regex
     not matching when datalad is run under argcomplete (for reasons I
     haven't looked into).  But, when I worked around that error,
     _then_ fb38c7098 broke completion.  Note that the cfg_kv_regex
     matching has since changed (b733b0472, 2020-04-17) in a way that
     avoids the config.py exception.