ENH: Overall progressbar for AnnexJsonProtocol
Until now it would only create individual progress bars for items that
git-annex would report progress about. This change adds an additional
global progress bar that counts the number of results, if a total number
of expected results was provided.

This fixes gh-4501

Or rather, it fixes the underlying desire to see something happen, it
won't change how and when results are emitted.