TST: addurls: Avoid leakage between two subsequent checks
test_addurls_unbound_dataset() looks at relative path handling, and
one of its checks asserts that the expected files exist.  But the same
dataset is used across two calls to addurls(), which means the check
for the second call could pass due to the previous call.

Use a separate dataset for each call.