BF: addurls: Don't trip over "./" in the file name format
As described in d46846716 (DOC: addurls: Clarify what filename format
is relative to, 2019-08-02) and the tests, the file format is intended
to be taken relative to the dataset.  However, when the file name
format begins with "./", we keep "./" on the relative name, while
constructing the absolute file name without the subdirectory as
documented.  The relative file name is passed to AnnexRepo.addurl(),
causing the file to be treated as relative to the current working
directory and downloaded there.

Achieve the documented behavior by applying the same handling to the
top-level dataset as we use for subdatasets.

Note that there is mention of changing this relative path handling in
gh-4498.  Even if we decide that this change in behavior is
worthwhile, it still makes sense to fix this bug, and doing so doesn't
make the future change harder.