TST: archive: Skip tests that depend on p7zip
All of the checks of the test generator fail on a system without
p7zip.  Before gh-4041 (in particular, before beb5d045e), the subset
of checks that existed at that time was skipped with (see gh-3176):

    SKIP: cmd:7z is missing. (Not) Funny enough but ATM we need p7zip
    installation to handle .gz files extraction 'correctly'

However, with the change to the archive file name in 05e9353d7 (TST:
Expand test coverage to additional relevant compression formats,
2020-01-17), the MissingExternalDependency exception that led to the
skip for gzip formats is no longer triggered.  And even if it were,
the other file formats tested would fail.

Mark these tests as known failures when p7zip isn't available.