TST: create-sibling: Drop kludge for git versions below 2.4
When checking the git version, test_create_sibling.py assumes that the
SSH target's git is the same as the local one, which doesn't work if a
Docker container is used as the SSH target.  We could instead use
SSHConnection.get_git_version(), but the git version kludge checks for
v2.4; that's old enough to not worry about because the tests will hit
other errors with git versions below v2.13.0 (see ff042b88a).