TST: annexrepo: Clear repo instances when testing shared connections
test_annex_ssh() goes through a convoluted dance to create an annex
repo, set up an SSH remote through GitRepo.add_remote(), and then get
an AnnexRepo instance.  The idea seems to be to add a remote without
using AnnexRepo.add_remote(), because that calls
_set_shared_connection(), and then check that AnnexRepo.__init__()
sets the shared connect.

However, __init__() isn't actually invoked because there's already a
known instance, and that's returned instead.  Clear the known instance
so that __init__() is triggered.

This appears to be a necessary part of the hang from gh-4288, though
there was presumably also a change on git-annex's end that exposed the
issue.  Here's a Travis test run from the parent commit (plus some
.travis.yml tweaks on top) that hangs:

  https://travis-ci.org/github/datalad/datalad/builds/705881743

And here's a complete run with the change here that does not hang:

  https://travis-ci.org/github/datalad/datalad/jobs/705850665