TST: create-sibling: Skip permissions test when running under sudo
test_failon_no_permissions() is failing on master's 'sudo -E' cron
build (*) because, not surprisingly, the "no permissions" error that
the test looks for is not raised.  The entire purpose of this test
isn't relevant when running under sudo, so skip it.

Note that this failure is specific to master.  Before 154e22281 (TST:
create-sibling: Extend some existing tests to test local paths,
2020-02-24), the test was restricted to an SSH target, and all
SSH-based tests are already disabled for the 'sudo -E' build.

(*) https://travis-ci.org/github/datalad/datalad/jobs/684799579#L1117