ENH: reenable testing using httpretty (and other tests)
Apparently we had the entire test_http disabled due to httpretty not supporting
python3 in the past.  That masked some bugs in our code which need to be fixed up
(done in prior commit) and minimal compatible version of httpretty
verified.

0.9.4 seems to work fine for me in conda environments with python 3.6
and 3.8.

The only leftover inconvenience (does not cause failed tests) is a
complaint from cookies upon exit:
https://github.com/datalad/datalad/issues/4542