ENH: Delay query of `get_content_annex_info()` in `push()`
Previously this was done in the initial `diff()`. This is often not
needed (e.g. --force no-datatransfer), and it is also too expensive,
because diff will obtain annex info twice (for 'from' and 'to' state of
the diff, while `push()` only cares about the 'to' state).

Now it is done only once, and at the latest possible stage.

The change include a TODO note with further optimization possibilities
that need further investigation.