ENH: Disable full submodule evaluation for diff(to=!None)
In all such cases the comparison is made with respect to a recorded
state, so that any worktree inspection can be completely avoided.
This can lead to substantial speed-ups in large dataset hierarchies,
and helps to mitigate the needs for alternatively tailored `diff`
implementations, such as the one proposed in gh-4507.