BF: use '' as default (not None) while checking for a key not to be a test key
I just saw in https://github.com/datalad/datalad-extensions/runs/665491656?check_suite_focus=true
which I believe failed for other reasons (but may be not), that there was a traceback
to this line.  And indeed above we might ask for a 'key' or a 'file', so hypothetically
there might be no 'key' field.  I have not really analyzed situation on how we have not
hit it before and how it really emerges in that failure while testing against
bleeding edge git annex