RF: throw dedicated RuntimeError with information about question in case of SilentConsoleLog
It is hard to impossible to troubleshoot failures of datalad in automated
deployments where we want to ask some question for some reason, but UI simply
does not support that since it is rightfully silent/non-interactive.

See e.g. https://github.com/templateflow/templateflow/issues/50 .

With this PR we would still fail but with a little more informative RuntimeError,
instead of obscure AttributeError stating what was actually asked.  We will not
disclose possible present other fields other than question and title if it was a
question asking for secret information ("hidden").  AFAIK in our use of the dialogues
we do not disclose any sensitive information in title or the question but I could be
proven wrong.

I did position this one on top of maint since might benefit people using released
version in production.  Delaying till 0.13.0 would require them to migrate first to
just be able to troubleshoot