BF: Prevent ugly interaction or result and progress reporting
This is an excerpt from gh-4444 that solely addresses the suboptimal
interaction of result reporting and progress reporting that is presently
evident in `push()` usage.

The idea is to clear any progressbar(s) before rendering a result with
`ui.message()` and afterward refresh the progressbar(s). With this
approach results will appear above any active progress bar in the
terminal and do not get half-swallowed by a progress update.

https://github.com/datalad/datalad/pull/4444#issuecomment-620285637
outlines an additional improvement that addresses the analog issue
for the interaction of log messages and progress reporting, also
pointed out by (but unaddressed in) gh-4444. However, that solution
can be seen as an incremental step that should not delay the fix
contained herein.