TST: sshconnector: Loosen assertion about SSH annex version
test_ssh_git_props() asserts that the ssh://localhost git-annex
version is equal to the version in external_versions.  However, those
versions don't necessarily match.  For example, the local user could
be running a git-annex from conda while the git-annex used by
ssh://localhost is from NeuroDebian.  And if in the future we move
away from a localhost target (e.g., gh-4514), there is no reason to
expect the versions to match at all.

Replace the more specific version assertion with a check that some
output is produced, as is already done for the SSH git version.

Fixes one part of gh-4554.