TST: sshrun: Don't fail if target doesn't accept LC_* variables
As mentioned in the comment, this test assumes the target has
`AcceptEnv LC_*` configured.  Instead of failing when that's not the
case (or dropping the test entirely), skip the test if the output is
an empty string, which probably means `AcceptEnv LC_*` isn't set.  (It
could also be an indication that things aren't wired up correctly in
sshrun, but this probably isn't worth worrying about given our test
environments have are configured to accept LC_ variables.)

Fixes one part of gh-4554.