BF: publish: Tell git-fetch to not recurse into submodules
publish() fetches the sibling before pushing, and has done so for a
long time (f07fc5194, 2017-08-26).  If the fetch brings down a
recorded submodule commit that is not yet present locally, by default
git also fetch in the submodule with the missing commit.

This is problematic because, as explained in dbd84e5662 (ENH: update:
Try to fetch submodule commit explicitly if needed, 2020-02-07),
submodule.c hard codes 'origin' as a fallback to fetch from, so the
fetch will fail unless there's a remote named 'origin'.  Also,
conceptually, the goal of this fetch seems to be to obtain up-to-date
information on the current repository, not its submodules, which
publish() will handle itself if called with --recursive.

Fetch with --recurse-submodules=no.