MNT: CHANGELOG.md: Fix incorrectly copied entry
I wrongly thought I had this in master's changelog (since this commit
was cherry picked) and copied over the first entry that looked
relevant.  But that was for an old issue.  Based on the attached notes
on 938d82097 (MNT: CHANGELOG.md: Fourth batch for 0.13.0, 2020-03-25),
it looks like I decided not to include an entry.

Anyway, include a correct description of 4332 now.

[ci skip]