TST: Add tests for ORA-HTTP
Note, that for now HTTP support is read-only and the we test the special
remote herein, not the server side. Since ORA just sends HEAD/GET
requests to the annex/objects/dir/hash/key/key of the dataset in the RIA
store, it's sufficient to serve those key there for testing. Archive
access and layout change to dirhashlower has to happen server side and
be delivered as response to those same requests.