RF: Add explicit config flag to enable `annex copy --auto`
This is the implementation of the conclusion of a discussion on this
topic. Follow-up to gh-4541.

Add --auto to `git annex copy` if `datalad.push.copy-auto-if-wanted`
is set, and `--force datatransfer` isn't given.

Fixes gh-4550
  --auto is never added, unless requested via config

Fixes gh-4551 and fixes gh-4483
  now obeys distribution-restrictions in a "default"
  `datalad push --to target` if configuration says so
  (not the default (yet?))

TODO:

Is this secure enough (see gh-4357). A dataset could set this
flag in its own config (.datalad/config). Are we confident that
it would never override a user config (add test)?