NF: dataset cache for tests
This introduces two test decorators for caching remote datasets for
tests. The caching location defaults to user level cache dir for datalad
under tests/ (GNU/Linux: $HOME/.cache/datalad/tests) and can be
configured via 'datalad.tests.cache'. Not specifying this, disables
caching altogether.
The intend is for a transparent cache, so decorated tests do not know
whether or not caching is enabled and must not depend on it. The dataset
in the cache is a drop-in replacement for the one at the specified URL.