DOC: manual: Warn about distribution-restrictions interactions
The metadata section of the manual describes using the
distribution-restrictions field to mark files that shouldn't be shared
publicly. As described at gh-4605, this setup depends on the caller
not configuring numcopies to a value above the default of 1 (or, less
likely, configuring required content in a way where the public remote
would want the file).