ENH: Give clone candidates a priority
This is achieved by simply prefixing their existing label with two digits
and sorting them before processing them.

This implementation makes no change to the previously established order,
it just makes that order explicit in the labels.

The default setup uses the priorities in the second half of the
spectrum:

50 for remote URL + submodule path
60 for the configured submodule URL
70 for any unprioritized 'datalad.get.subdataset-source-candidate' config
90 for the local subdataset path

The first half of the spectrum and the spaces inbetween are available for
re-prioritizing clone attempts. The idea is to be able to configure

datalad.get.subdataset-source-candidate-00takemefirst = http://some

and have this be processed first. Likewise using '99label' it is
possible to configure a fallback source candidate.