RF: --transfer-data=auto-if-wanted which dissolves datalad.push.copy-auto-if-wanted config
I was going back and forth either we need some additional "datalad.push.transfer-data"
config setting which could be set per repository to indicate what "datalad push"
should do with data.  But I remain of an idea that it would be counter-productive
to establish yet another level of data transfer behavior instead of using what is
already there provided by git-annex itself.

To maintain "do the right thing" behavior by default I had decided also to make
auto-if-wanted to be the default mode of operation -- it will transfer data if there
is no explicit "wanted" setting for the remote.