TST: publish: Adjust check for an auto-init'ing git-annex
test_publish_simple() and test_publish_recursive() have started to
fail in the datalad-extensions builds that test against an unreleased
git-annex [1].  The failing checks assert that the git-annex branch in
the source and target repos has the same set of commits.  This is no
longer true as of git-annex's 484a74f07 (auto-init autoenable=yes,
2020-05-27) because the target gets initialized by git-annex, which
means the git-annex branch includes a change to uuid.log.

gh-1327 dealt with essentially the same issue for a downstream check.
Apply the same solution to these other spots: assert only that tip of
the git-annex branch from the source repo is contained within the
target's git-annex branch.

Closes #4618.

[1] example: https://github.com/datalad/datalad-extensions/runs/750780790