TST: publish: Check git-annex branch consistently across tests
There have been a few git-annex branch checks in test_publish.py [1]
that have started to fail with newer git-annex versions.  All of these
assume a specific set of git-annex commits on the target, which is
making too strong of an assumption about git-annex's behavior.  The
safer check, and the only one that should really matter to publish(),
is that the state of the source's git-annex branch got incorporated
into the target's.

To hopefully avoid future failures like this, add an explicit helper
for this check and use it in all of the remaining spots in
test_publish.sh that check that the source's git-annex was published
to the target's.

It probably makes sense to use this in test_push.py as well, though
none of those test fail with the development git-annex (likely because
push isn't coupled with create-sibling).  Punt on that since it's
trivial to extract this helper later.

[1] one way back in gh-1319 and two recently in gh-4618, addressed by
    the last commit