RF: create_sibling: Unconditionally set HEAD for sibling
create-sibling sets the sibling's HEAD to the current branch name if
it's not master.  This is the only spot outside the tests that hard
codes master as the name of the default branch.

To work with Git versions that allow customizing the default branch
name, we could update the check to query for it.  But given that the
code is unnecessary but not invalid for the default branch, let's just
always perform the call.