TST: Replace all hard-coded uses of master
As of Git 2.28, it is possible to configure the default branch used
when creating a repository, and some version after that may replace
master with another name.

Starting a few commits back, setup_package() now configures the
default branch name to a custom value to test that DataLad
functionality does not rely on a hard-coded default branch name.