TST: tests.utils: Override init.defaultBranch to custom value
As of Git 2.28.0, the unborn branch in a newly created repository is
named according to init.defaultBranch, falling back to master.  A
later phase (perhaps in the time frame of three major releases [1])
may change the default branch name to something other than master.

Functionally our code base already seems to be in good shape with
respect to that change, aside from tests.  Many spots of the tests
assume that master is the default branch of the repository (regardless
of whether they avoid assuming it is the current branch for
compatibility with adjusted branches).

Until the default changes, tests that hard code master will still pass
in most cases because setup_package() overrides HOME and uses custom
Git configuration unless GIT_HOME is set.  That means users setting
init.defaultBranch globally in their ~/.gitconfig would not have an
effect.

Still, testing with a branch name other than master is useful to (1)
hopefully catch issues that users would hit into in the wild if they
set init.defaultBranch and (2) future-proof our tests for changes in
the default.

[0] https://lore.kernel.org/git/pull.656.v4.git.1593009996.gitgitgadget@gmail.com/
[1] https://lore.kernel.org/git/xmqqeeqiztpq.fsf@gitster.c.googlers.com