DOC: annexrepo: Add note about issue with older Git versions
In tests with Git v2.9 (which is the oldest Git version that will
build for me on Debian Buster), test_AnnexRepo_file_has_content()
fails:

    File "[...]/datalad/support/tests/test_annexrepo.py", line 318, in
     test_AnnexRepo_file_has_content [ar.supports_unlocked_pointers])
    AssertionError: [False] != [True]

The failure is coming from the 'git diff' call unexpectedly reporting
a staged type change as modified.  Adding a 'git status' call before
the 'git diff' call is enough to convince Git that there is no
modification, so perhaps it's an issue related to not triggering the
clean filter.

This failure disappears with Git v2.13.