BF: progressbar: Add reset() kludge for older tqdm versions
As of c9a066c49 (ENH: Enable changing the total of an existing
progress bar, 2020-04-22), our tqdmProgressBar.update() relies on tqdm
having reset(), but reset() didn't exist until v4.32.0 (c7015f4,
2019-05-10).  Debian Buster currently has v4.28.1.

For compatibility with older versions, copy over the logic to
tqdmProgressBar and fall back to it when tqdm.reset() isn't available.
All of the attributes adjusted in reset() appear to be in v4.28.1 and
manual testing with v4.28.1 didn't reveal any issues.

There is of course some older version that this won't work for, but it
looks like we'll hit other problems before that point.  We rely on
refresh() throughout tqdmProgressBar, and all of the other attributes
used in _reset_kludge() appear to have been around before refresh()
was added (d3c3ad2, 2016-01-30).

Closes #4648.