ENH: gitrepo: Declare and check minimum Git version
We do not specify a minimum supported Git version, which makes it
difficult for us to decide whether we can use a newer feature and
means that users have to rely on trial and error to know whether their
version is recent enough.

Set a minimum Git version, handling it in a way that closely follows
what AnnexRepo already does for the minimum git-annex version.  Choose
2.19.1 (September 2018) as the minimum because that (1) matches what
conda currently has for its 32-bit Linux builds [0] and (2) is below
what is currently in Debian stable (v2.20.1).

Note that if we decide to lower the minimum more, the hard floor
should probably be considered v2.13.0 (May 2017).  As described a few
commits back, AnnexRepo._check_files() doesn't work reliably until
then and testing with v2.13.0 didn't reveal major problems [1], though
the skips added in the two previous commits were needed.

[0] https://github.com/datalad/datalad/issues/4636#issuecomment-645633006
[1] https://travis-ci.org/github/datalad/datalad/builds/701067382

Closes #4636.