BF: get: Inherit reckless configuration when cloning subdatasets
Cloning a subdataset is supposed to inherit the parent's
`datalad.clone.reckless` value.  This works as intended when using
install() to get an existing dataset or using clone() to create a new
subdataset, but the parent's configuration is ignored when the clone
is triggered by get().

The "inherit reckless" logic is handled in Clone.__call__(), but
that's upstream of clone_dataset(), where get() hooks into.  Move the
handling into clone_dataset() to avoid repeating it in get().

Note that we don't have access to the parent dataset in
clone_dataset(), but we need to query its configuration.  We could add
the parent dataset as a parameter.  However, there's already a cfg
parameter, and Clone.__call__()---the only caller that specifies
cfg---passes the parent's configuration.  So instead just update get()
to pass in cfg and query that, adjusting the documentation to note
that cfg should be the parent's.

Closes #4655.