TST: travis: Add job for python 3.8
Before Python 3.8 was released, we tested against it and fixed up the
issues at the time, but it wasn't added as a regular job.  Add it now,
demoting the 3.6 job to a cron job so we don't increase our testing
time.

Use Bionic rather than Xenial as the distribution so that we can use a
pre-installed 3.8 instead of 3.8-dev [0].  Note that we can't use this
across all runs because Bionic doesn't come with Python 3.5
pre-installed.

[0] https://docs.travis-ci.com/user/reference/bionic/#python-support