ENH: Let default result renderer tolerate non-path results
Example use case: credential query
(https://github.com/datalad/datalad-osf/pull/105)

There is nothing on the file system that such credentials are (or have
to be) associated with.

Before this change, the default result handler would crash on results
without a `path` key. After this change, it merely doesn't report a
path at all. Spacing is adjusted be

   `<action>(<status>): [<message>]`

in this case.