RF: move instantiation of TMPDIR before installing git annex
Which gets me thinking that may be we should not just export TMPDIR straight
in the environment but rather e.g. as TMPDIR_ and then assign to TMPDIR after
they are setup, so git annex installation does not end up on some odd TMPDIR
to be tested.  But I will commit this to see what effects this might have.

Also it reminded about a workaround for git-annex we established via config
-- hardcoded pidlock setting.  git-annex should figure it out on its own!