TST: with_sameas_remote: Skip on older systems
Our testing setup with sameas rsync/dir special remotes hangs on
Travis [0].  This appears to be an interaction of a recent git-annex
commit, 1f2e2d15e (async exception safety, 2020-06-03), and an older
openssh or rsync version [1].

When running with newer git-annex versions, make with_sameas_remote()
helper skip the test if openssh is below 7.4 because 7.4p is on Debian
Stretch and doesn't have a hang, while Debian Jessie with 6.7p1 does.
This is crude (e.g., maybe rsync is the problem instead of openssh),
but it like is a good enough indicator of "too old" for our purposes.
Note that, even if we start testing with a newer git-annex and it
doesn't get a fix for this, we can still make sure to cover these
skipped test on Travis by keeping a Bionic run.  (We currently have
one.)

Along with the recent fix in git-annex for handling SSH options [2],
this resolved all the hangs on Travis when running with a recent
git-annex (8.20200618-gc9d0bf0e6) [3].

Closes #4499.

[0] https://travis-ci.org/github/datalad/datalad/builds/704464417
[1] https://git-annex.branchable.com/bugs/Recent_hang_with_rsync_remote_with_older_systems___40__Xenial__44___Jessie__41__/
[2] f912f8e5f (refix bug in a better way, 2020-07-02)
    https://git-annex.branchable.com/bugs/annex-ssh-options_dropped_since_8.20200330/
[3] https://travis-ci.org/github/datalad/datalad/builds/705920487