BF: get_content_info: Account for change in ls-tree output
get_content_info() relies on ls-tree showing an entry for a submodule
when given a path _within_ the submodule:

    $ git ls-tree HEAD sub/file
    160000 commit 05b9a93e327b959cd6ead46debf5f67cabab3a2a	sub

However, that longstanding behavior appears to be unintended
consequence of a fix to make 'ls-tree sub' and 'ls-tree sub/' return
the same result.  As of Git's 35a9f1e99c (tree-walk.c: don't match
submodule entries for 'submod/anything', 2020-06-05), the output for
the above command is empty.

That commit makes ls-tree match ls-files in how it treats paths within
submodules [*].  get_content_info() uses ls-tree when ref=None, so it
already has logic to map from sub-paths to submodules.  Reposition
that logic outside so that it can also be applied to paths destined
for ls-tree.

Closes #4681.

[*] ... resolving the inconsistency that @yarikoptic reported at
    <https://lore.kernel.org/git/20190703193305.GF21553@hopa.kiewit.dartmouth.edu>,
    though in the direction that's inconvenient for our purposes.